% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_frip_dt}
\alias{plot_frip_dt}
\title{plot_frip_dt}
\usage{
plot_frip_dt(
  frip_dt,
  peak_dt = NULL,
  fq_dt = NULL,
  query_gr = NULL,
  sort_by = c("frip", "N")[1],
  name_lev = NULL
)
}
\arguments{
\item{frip_dt}{output from \code{\link{make_frip_dt}}}

\item{peak_dt}{(optional) output from \code{\link{make_peak_dt}}}

\item{fq_dt}{(optional) output from \code{\link{make_fq_dt}}}

\item{query_gr}{(optional) The GRanges used to create frip_dt, only necessary to calculate the fraction of genome covered by peaks.}

\item{sort_by}{character. One of "frip" or "N". Should plots be sorted by decreasing FRIP ("frip") or total reads ("N")? Default is "frip".}

\item{name_lev}{character. Name levels to impose manual ordering.  sort_by is not ignored if this is not NULL. Default is NULL.}
}
\value{
list of ggplot plots relevant to FRIP.
}
\description{
plot_frip_dt
}
\examples{
peak_file = dir(system.file("extdata", package = "seqqc"), pattern = "test_peaks.bed$", full.names = TRUE)
bam_file = dir(system.file("extdata", package = "seqqc"), pattern = "test_peaks.bam$", full.names = TRUE)

query_gr = easyLoad_bed(peak_file)[[1]]

frip_dt = make_frip_dt(bam_file, query_gr)
}
