% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_write.R
\name{write_bed_frip}
\alias{write_bed_frip}
\title{write_bed_frip}
\usage{
write_bed_frip(query_gr, assign_dt, frip_dt, file = "regions_with_FRIP.txt")
}
\arguments{
\item{query_gr}{GRanges of regions used to make frip_dt and assign_dt}

\item{assign_dt}{data.table with "id" and "cluster_id" that maps region ids to clusters. In output from \code{\link{plot_signals}}.}

\item{frip_dt}{output from \code{\link{make_frip_dt}}}

\item{file}{File to write to}
}
\value{

}
\description{
write_bed_frip
}
