% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_write.R
\name{write_bed_overlaps}
\alias{write_bed_overlaps}
\title{write_bed_overlaps}
\usage{
write_bed_overlaps(query_gr, assign_dt, file = "regions_with_overlaps.txt")
}
\arguments{
\item{query_gr}{GRanges of regions used to make assign_dt}

\item{assign_dt}{data.table with "id" and "cluster_id" that maps region ids to clusters. In output from \code{\link{plot_signals}}.}

\item{file}{File to write to}
}
\value{

}
\description{
write_bed_overlaps
}
