% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_signals}
\alias{plot_signals}
\title{plot_signals}
\usage{
plot_signals(
  prof_dt,
  query_gr,
  anno_dt = NULL,
  frip_dt = NULL,
  name_lev = NULL
)
}
\arguments{
\item{prof_dt}{data.table of signal profiles from \link[seqsetvis]{ssvFetchBam} or \link[seqsetvis]{ssvFetchBigwig}.}

\item{query_gr}{The GRanges used to produce prof_dt.}

\item{anno_dt}{(optional) output from \code{\link{make_anno_dt}}}

\item{frip_dt}{(optional) output from \code{\link{make_frip_dt}}}

\item{name_lev}{(optional) manual specification for levels of "name" to control ordering.}
}
\value{
named list of ggplot plots and relevant data
}
\description{
Performs clustering on signal profiles in prof_dt and produces various plots
}
\examples{
bw_files = dir(system.file("extdata", package = "seqqc"), pattern = "bw$", full.names = TRUE)
query_dt = make_dt(bw_files)
query_dt[, sample := sub("_FE_random100.A", "", name)]

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = easyLoad_narrowPeak(peak_files)
query_gr = resize(ssvOverlapIntervalSets(peak_grs), 6e2, fix = "center")
anno_dt = make_feature_as_signal_dt(anno_grs, query_gr)

prof_dt = ssvFetchBigwig(query_dt, query_gr, return_data.table = TRUE)

sig_res = plot_signals(prof_dt, query_gr)
sig_res$heatmap

sig_res.anno = plot_signals(prof_dt, query_gr, anno_dt = anno_dt)
}
